feat(pic16f87xa-hal): MPLAB X / XC8 project template
Adds a minimal real-target build configuration:

- mcu/pic16f87xa-mplabx/Makefile — produces build/<MCU>-firmware.hex
  via xc8-cc. Picks up the right HAL source list per device
  (PSP source is added only for 40/44-pin parts). Emits the
  Configuration Word (§14.1) into an auto-generated .c file:
  FOSC=HS, WDTE=ON, PWRTE=ON, BOREN=ON, LVP=OFF, WRT=OFF.

- mcu/pic16f87xa-mplabx/nbproject/configurations.xml — IDE-format
  standalone-project descriptor. References all HAL headers
  and sources plus tests/example_blink.c as the application
  entry point.

- mcu/pic16f87xa-mplabx/README.md — explains how to open the
  project in MPLAB X, and (if the IDE balks) how to create a
  fresh project and add the existing sources.

The CMake build (host simulation) and the Makefile build
(real silicon) are completely orthogonal: PIC16F87XA_USE_SIMULATOR
selects between them, and the sim/ directory is never compiled
on real targets.

+++ pic16f87xa-hal/tests/example_eeprom.c
@@ -25,7 +25,7 @@
 } while (0)
 
 /* Helper: read a register from a non-default bank. */
-static uint8_t b_read(uint8_t bank, uint8_t addr)
+static uint8_t b_read(uint8_t bank, uint16_t addr)
 {
     uint8_t prev = (PIC16F87XA_REG8(PIC_REG_STATUS) >> 5) & 0x03U;
     pic_select_bank(bank);
